Break up IncrementalCompilationState.swift to protect the protectedState variable
